# docs/api-reference/swagger.yaml
openapi: "3.0.3"
info:
  description: Sandbox API
  version: 1.0.0
  title: Sandbox API
  license:
    name: MIT
servers:
  - url: /api/v1
...
paths:
...
  /placements/{uuid}:
    parameters:
      - in: header
        name: Authorization
        description: Access JTW Token
        required: true
        schema:
          type: string
        example: Bearer <ACCESS_TOKEN>
      - name: uuid
        in: path
        required: true
        description: The UUID of the service.
        schema:
          $ref: "#/components/schemas/UUID"
    get:
      tags:
        - placement
      operationId: getPlacement
      summary: Get information about a placement and its resources
      responses:
        '200':
          description: The placement assigned to this uuid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Placement"
              example:
                service_uuid: 6548dc97-5799-4bfe-8843-8d1793996593
                status: success
                resources: []
                annotations:
                  guid: testguid
        '404':
          description: getPlacement Placement not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: getPlacement unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - placement
      operationId: deletePlacement
      summary: Delete the placement assigned to a specific uuid.
      description: |-
        You usually call this endpoint once your service is destroyed.

        This action will cascade cleanup for all the accounts associated with the placement.
      responses:
        '202':
          description: The placement assigned to this uuid is deleted. The resources associated with this placement are marked for cleanup.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                message: Placement marked for deletion

        '404':
          description: deletePlacement Placement not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: placement not found
                http_code: 404
        default:
          description: deletePlacement unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /accounts/{kind}/{name}/cleanup:
    parameters:
      - in: header
        name: Authorization
        description: Access JTW Token
        required: true
        schema:
          type: string
        example: Bearer <ACCESS_TOKEN>
      - name: kind
        in: path
        required: true
        description: The kind of account
        schema:
          type: string
          enum:
            - aws
            - AwsSandbox
            - ocp
            - OcpSandbox
            - dns
            - DNSSandbox
            - ibmrg
            - IBMResourceGroupSandbox
            - ssl
            - SSLSandbox
      - name: name
        in: path
        required: true
        description: The name of the sandbox to mark for cleanup.
        schema:
          type: string
    put:
      tags:
        - account
      operationId: markForCleanup
      summary: Manually mark a sandbox (account) for cleanup.
      description: |-
        This is an ad-hoc operation, deleting placement should be preferred for regular cleanup of sandboxes.
        If the sandbox is part of a placement, the corresponding placement will be updated:
          - The sandbox is removed from the placement.

        The placement is automatically deleted if there is no more resources.
      responses:
        '200':
          description: The account is marked for cleanup.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: markForCleanup unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    UUID:
      type: string
      description: The UUID of the placement. Sandboxes are assigned uuid when booked. Usually anarchy UUID.
      format: uuid
      example: 6548dc97-5799-4bfe-8843-8d1793996593
...
    BasePlacement:
      description: Base placement fields shared by placement and dry-run requests
      type: object
      properties:
        reservation:
          type: string
        resources:
          type: array
          items:
            $ref: '#/components/schemas/ResourceRequest'
        annotations:
          $ref: '#/components/schemas/Annotations'
      required:
        - resources
    ResourceRequest:
      description: A single resource request within a placement
      type: object
      properties:
        kind:
          $ref: '#/components/schemas/ResourceKind'
        count:
          type: integer
          minimum: 1
          default: 1
          example: 1
        alias:
          type: string
          description: |
            A label for this resource within the placement request.
            Used by `cluster_condition` in sibling resources to reference this resource.
            Must be unique within the placement request.
          example: cluster
        cluster_condition:
          type: string
          description: |
            An expression that constrains cluster scheduling relative to sibling resources.
            Uses `same(<alias>)`, `different(<alias>)`, or `child(<alias>)` functions,
            combined with `&&` (AND).

            - `same(A)`: schedule on the same cluster as the resource with alias `A`
            - `different(A)`: schedule on a different cluster than alias `A`
            - `child(A)`: schedule on a child (HCP) cluster of the cluster used by alias `A`

            All referenced aliases must exist in the placement request.
            Resources with `cluster_condition` wait for their referenced siblings
            to be provisioned before scheduling begins.
          example: "same(cluster) && different(other)"
        cluster_relation:
          type: array
          description: |
            Explicit cluster relation constraints (alternative to `cluster_condition`).
            Prefer using `cluster_condition` for new integrations.
          items:
            type: object
            properties:
              relation:
                type: string
                enum:
                  - same
                  - different
                  - child
              reference:
                type: string
        cloud_selector:
          description: |
            Annotations used to match clusters. All key-value pairs must match
            the cluster's annotations (AND logic across keys).

            Values may contain pipe-separated alternatives for OR logic within
            a single key. For example, `purpose: prod|event` matches clusters
            whose `purpose` annotation is either `prod` or `event`.

            When multiple keys use pipes, all combinations are tried (cartesian
            product). For example:
            ```yaml
            cloud_selector:
              purpose: prod|event
              cpuType: amd|intel
            ```
            matches clusters with (`purpose=prod` AND `cpuType=amd`) OR
            (`purpose=prod` AND `cpuType=intel`) OR (`purpose=event` AND
            `cpuType=amd`) OR (`purpose=event` AND `cpuType=intel`).
          allOf:
            - $ref: "#/components/schemas/Annotations"
        cloud_preference:
          $ref: "#/components/schemas/Annotations"
        keycloak_user_prefix:
          type: string
          description: |
            Prefix for the Keycloak username created when `keycloak` is enabled in `cloud_selector`.
            The resulting username will be `<prefix><guid>`.
            Defaults to `sandbox-` if not specified.
          example: "sandbox-"
        no_namespace:
          type: boolean
          description: |
            When set to true, no namespace, service account, RBAC, quota, limit range,
            keycloak user, or HCP/virt resources will be created on the cluster.
            Only the cluster information (API URL, ingress domain, console URL) and the
            deployer-admin service account token are returned.
            This is incompatible with `keycloak: yes` in `cloud_selector`.
          default: false
          example: false
        annotations:
          $ref: "#/components/schemas/Annotations"
        quota:
          type: object
          description: |
            Quota to be applied to the namespace of the OcpSandbox
            The way quota is applied depends on the OcpSharedClusterConfiguration.
            By default, if a quota value is missing, the default quota of the shared cluster is applied instead.

            StrictDefaultSandboxQuota is a flag to determine if the default sandbox quota should be strictly enforced. If set to true, the default sandbox quota will be enforced as a hard limit. Requested quota not be allowed to exceed the default. By default it's false.

            If set to false, the default sandbox will be updated to the requested quota.
            QuotaRequired is a flag to determine if a quota is required in any request for an OcpSandbox.
            If set to true, a quota must be provided in the request.
            If set to false, a quota will be created based on the default sandbox quota.
            By default it's false.
          example:
            pods: "12"
            requests.memory: "20Gi"
            limits.memory: "20Gi"
            secrets: "2"
            requests.ephemeral-storage: "100Gi"
        limit_range:
          type: object
          description: |-
            Limit Range for the sandbox.
            This allows to set the default limit and request for pods.
            see https://kubernetes.io/docs/concepts/policy/limit-range/

            Two formats are accepted:

            **Full format** (standard Kubernetes LimitRange spec):
            ```
            limit_range:
              spec:
                limits:
                  - type: Container
                    default:
                      cpu: "1"
                      memory: 2Gi
                    defaultRequest:
                      cpu: "0.5"
                      memory: 1Gi
            ```

            **Shorthand format** (default and defaultRequest at the top level):
            ```
            limit_range:
              default:
                cpu: "1"
                memory: 2Gi
              defaultRequest:
                cpu: "0.5"
                memory: 1Gi
            ```

            The shorthand format is automatically converted to a full
            LimitRange with type "Container".
          example:
            default:
              cpu: "1"
              memory: 2Gi
            defaultRequest:
              cpu: "0.5"
              memory: 1Gi
    PlacementRequest:
      description: |-
        A placement request is a record of service(uuid) / resources / annotations
      type: object
      allOf:
        - $ref: '#/components/schemas/BasePlacement'
        - type: object
          required:
            - service_uuid
          properties:
            service_uuid:
              $ref: "#/components/schemas/UUID"
      example:
        service_uuid: 13a8b15c-e752-4727-ac78-600e8833e575
        resources:
          - kind: AwsSandbox
            count: 2
    Annotations:
      description: Key / Value map to provide optional information.
      type: object
      additionalProperties:
        type: string
        minLength: 1
      example:
        guid: abcde
        env_type: ocp4-cluster
        hcp: "yes"
...
    Placement:
      type: object
      description: A placement is a record of service(uuid) / resource(name) / when(date)
      required:
        - service_uuid
        - resources
        - annotations
      properties:
        request:
          $ref: "#/components/schemas/PlacementRequest"
        service_uuid:
          $ref: "#/components/schemas/UUID"
        resources:
          $ref: "#/components/schemas/Resources"
        status:
          type: string
          example: success
          enum:
            - new
            - initializing
            - updating
            - success
            - error
            - deleting
        created_at:
          description: The date (UTC and RFC3339 format) the placement was made.
          type: string
          format: date-time
          example: 2023-03-13T09:42:33+01:00
        updated_at:
          description: The date (UTC and RFC3339 format) the placement was last updated.
          type: string
          format: date-time
          example: 2023-03-13T09:42:33+01:00
        annotations:
          $ref: "#/components/schemas/Annotations"
...
    Account:
      description: A sandbox account, one of AwsAccount, AzureAccount, OcpSandbox, DNSSandbox, or IBMResourceGroupSandbox
      oneOf:
        - $ref: '#/components/schemas/AwsAccount'
        - $ref: '#/components/schemas/AzureAccount'
        - $ref: '#/components/schemas/OcpSandbox'
        - $ref: '#/components/schemas/DNSSandbox'
        - $ref: '#/components/schemas/IBMResourceGroupSandbox'
      discriminator:
        propertyName: kind
        mapping:
          AwsSandbox: '#/components/schemas/AwsAccount'
    Resource:
      description: A provisioned resource with availability and cleanup status
      type: object
      required:
        - available
        - to_cleanup
      properties:
        kind:
          $ref: '#/components/schemas/ResourceKind'
        to_cleanup:
          description: True if the resource is marked for cleanup
          type: boolean
          default: false
          example: false
        available:
          description: True if the resource has no placement associated.
          type: boolean
          default: true
          example: false
      oneOf:
        - $ref: '#/components/schemas/Account'
        - $ref: '#/components/schemas/AwsAccount'
      discriminator:
        propertyName: kind
        mapping:
          AwsSandbox: '#/components/schemas/AwsAccount'
...
    Resources:
      description: List of provisioned resources
      type: array
      nullable: true
      items:
        $ref: '#/components/schemas/Resource'
      example:
        - name: sandbox123
          kind: AwsSandbox
          account_id: '0123456'
          zone: sandbox123.domain.com
          hosted_zone_id: Z019612345AH4C7KTTZJ0
          available: true
          to_cleanup: false
...
    AzureAccount:
      description: An Azure sandbox account
      type: object
      properties:
        name:
          type: string
          example: sandbox123
...
    Error:
      description: Error response with HTTP code and message
      type: object
      required:
        - http_code
        - message
      properties:
        http_code:
          type: integer
          format: integer
          example: 502
        message:
          type: string
          example: Unexpected Error
    Message:
      description: Simple message response
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Account successfully marked for cleanup
...
    ResourceKind:
      description: The type of sandbox resource
      type: string
      example: AwsSandbox
      default: AwsSandbox
      enum:
        - AwsSandbox
        - OcpSandbox
        - DNSSandbox
        - IBMResourceGroupSandbox
        - SSLSandbox
...